Issue #762: Cloud Pipeline shall allow users to configure spending quotas - review fixes

+++ api/src/main/java/com/epam/pipeline/entity/quota/QuotaActionEntity.java
@@ -49,7 +49,7 @@ public class QuotaActionEntity {
     @JoinColumn(name = "quota_id", referencedColumnName = "id")
     private QuotaEntity quota;
 
-    private Integer threshold;
+    private Double threshold;
 
     @Convert(converter = QuotaActionTypeListConverter.class)
     private List<QuotaActionType> actions;
